@W-22973720: Rename listDatasources/ dir to datasources/
Address PR review (r3414744028): mirror workbooks/ layout. Updated external refs; internal ./ imports unaffected.
